requirements.txt
flask==3.1.1
requests==2.32.4
beautifulsoup4==4.13.4
selenium==4.33.0
lxml==5.4.0
langchain>=0.3.0
langchain_groq>=0.1.0
langchain_community
python-dotenv
readability-lxml 